+++ src/main/java/lotto/Model/Cpu.java
@@ -1,9 +1,10 @@
-package lotto;
+package lotto.Model;
 
 import java.util.ArrayList;
 import java.util.List;
 import lotto.Constants.ErrorMessage;
 import lotto.Constants.IntConstants;
+import lotto.Controller.LottoGenerator;
 
 public class Cpu {
     private final List<Lotto> sixLottoNumbers;
