+++ src/main/java/lotto/Model/Lotto.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.Model;
 
 import java.util.HashSet;
 import java.util.List;
