+++ src/main/java/lotto/Model/Player.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.Model;
 
 import java.util.List;
 import lotto.Constants.ErrorMessage;
