+++ src/test/java/lotto/GameRuleTest.java
@@ -4,12 +4,13 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 import java.util.ArrayList;
 import java.util.List;
+import lotto.Controller.LottoCounter;
+import lotto.Model.GameRule;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 public class GameRuleTest {
     GameRule gameRule = new GameRule();
-    LottoCounter lottoCounter = new LottoCounter();
 
     @DisplayName("두 리스트를 비교해서 겹치는 숫자의 갯수를 반환하는 테스트입니다.")
     @Test
@@ -17,7 +18,7 @@ public class GameRuleTest {
         List<Integer> list1 = new ArrayList<>(List.of(1, 2, 3, 4, 5, 6));
         List<Integer> list2 = new ArrayList<>(List.of(1, 2, 3, 4, 5, 6));
 
-        int result = lottoCounter.countElement(list1, list2);
+        int result = LottoCounter.countElement(list1, list2);
 
         assertThat(result).isEqualTo(6);
     }
