+++ src/test/java/lotto/LottoCounterTest.java
@@ -4,13 +4,13 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.ArrayList;
 import java.util.List;
+import lotto.Controller.LottoCounter;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 
 public class LottoCounterTest {
 
-    LottoCounter lottoCounter = new LottoCounter();
 
     @Test
     @DisplayName("두 리스트에서 중복된 요소의 갯수를 찾는 메서드 테스트입니다.")
@@ -18,7 +18,7 @@ public class LottoCounterTest {
         List<Integer> list1 = new ArrayList<>(List.of(1, 2, 3, 4, 5, 6));
         List<Integer> list2 = new ArrayList<>(List.of(1, 2, 3, 7, 8, 9));
 
-        int result = lottoCounter.countElement(list1, list2);
+        int result = LottoCounter.countElement(list1, list2);
 
         assertThat(result).isEqualTo(3);
     }
